Revert "Hm3"
