--- src/transport/Bus.java
@@ -1,9 +1,0 @@
-package transport;
-
-public class Bus extends Transport{
-    public Bus(String brand, String model, String color, int year, String country, int maxSpeed) {
-        super(brand, model, color, year, country, maxSpeed);
-    }
-
-
-}
