--- src/transport/Transport.java
@@ -1,71 +1,0 @@
-package transport;
-
-public class Transport {
-    private final String brand;
-    private final String model;
-    private String color;
-    private final int year;
-    private final String country;
-    private int maxSpeed;
-
-    public Transport(String brand, String model, String color, int year, String country, int maxSpeed) {
-        this.brand = brand;
-        this.model = model;
-        setColor(color);
-        this.year = year;
-        this.country = country;
-        setMaxSpeed(maxSpeed);
-    }
-
-    public String getBrand() {
-        return brand;
-    }
-
-    public String getModel() {
-        return model;
-    }
-
-    public String getColor() {
-        return color;
-    }
-
-    public void setColor(String color) {
-        if(color == null || color.isEmpty()){
-            this.color = "Белый";
-        } else {
-            this.color = color;
-        }
-    }
-
-    public int getYear() {
-        return year;
-    }
-
-    public String getCountry() {
-        return country;
-    }
-
-    public int getMaxSpeed() {
-        return maxSpeed;
-    }
-
-    public void setMaxSpeed(int maxSpeed) {
-        if (maxSpeed <= 0) {
-            this.maxSpeed = 150;
-        }else {
-            this.maxSpeed = maxSpeed;
-        }
-    }
-
-    @Override
-    public String toString() {
-        return "Transport{" +
-                "brand='" + brand + '\'' +
-                ", model='" + model + '\'' +
-                ", color='" + color + '\'' +
-                ", year=" + year +
-                ", country='" + country + '\'' +
-                ", maxSpeed=" + maxSpeed +
-                '}';
-    }
-}
